Fixed the routing issues and move multer into the helper folder
