+++ backend/index.js
@@ -1,6 +1,5 @@
 const express = require("express");
-const app = require('./userSignUp');
-const { extractTextFromPdf } = require("./pdfExtractor");
+const userSignUp = require("./userSignUp");
 const jobDescriptionRoutes = require("./jobDescription")
 const resumeUploadRoutes = require("./resumeUpload");
 
@@ -8,6 +7,7 @@ const app = express();
 app.use(express.json());
 
 // Mount job description routes under /api
+app.use("/api", userSignUp);
 app.use("/api", jobDescriptionRoutes);
 app.use("/api", resumeUploadRoutes);
